Add ability to omit scanner-identifiers from conversion
Double Open's ORT workflow adds "LicenseRef-Scanner" to scaner findings,
so that they can be handled with different rules than concluded
licenses. This makes generation of notice files hard, as a notice text
would have to be added for example for "LicenseRef-Scanner-MIT". This
commit adds a flag to not add the scanner identifiers.

Signed-off-by: Mikko Murto <mikko.murto@hhpartners.fi>