fix(dos): Distinguish between requested and actual PURLs
Before this change, the provenance as determined by the downloader was
simply discarded, assuming that it was always exactly downloaded what
was requested. However, that is not correct in case of e.g. a moving
"master" VCS revision. It needs to be recorded which revision "master"
resolved to at the moment of downloading / scanning.

Also, the PURL itself did not contain any information about the origin of
the source code until now, which is required in order to disinguish a VCS
location from a source artifact location as these origins may contain
different source code.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>