fix(dos): Use the package's VCS path in the PURLs
Previously, PURLs were exclusively created from the provenance of a
package. However, the root provenance of a nested provenance has the VCS
path cleared to indicate that all covered packages stem from the exact
same source code origin.

But the PURLs created for DOS need to maintain the VCS path, even if it
points to the same source code origin, in order to get the "bookmarking"
right for the file tree in the package configuration UI. So mix and match
the required information by taking PURL qualifiers from the provenance
but the PURL subpath from the respective package, and create a PURL for
that.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>